requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
streamlit
requests
pandas